+++ src/main/java/fr/zcraft/quartzlib/tools/items/ItemUtils.java
@@ -37,7 +37,10 @@ import java.lang.reflect.Method;
 import java.lang.reflect.Type;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.Random;
+import org.bukkit.ChatColor;
+import org.bukkit.DyeColor;
 import org.bukkit.GameMode;
 import org.bukkit.Location;
 import org.bukkit.Material;
@@ -47,6 +50,9 @@ import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.ItemMeta;
 import org.bukkit.inventory.meta.PotionMeta;
 import org.bukkit.potion.Potion;
+import org.jetbrains.annotations.Contract;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 //import org.bukkit.Sound;
 
@@ -103,7 +109,7 @@ public abstract class ItemUtils {
      * @param player The player to give the item to.
      * @param item   The item to give to the player
      * @return true if the player received the item in its inventory, false if
-     *     it had to be totally or partially dropped on the ground.
+     *              it had to be totally or partially dropped on the ground.
      */
     public static boolean give(final Player player, final ItemStack item) {
         final Map<Integer, ItemStack> leftover = player.getInventory().addItem(item);
@@ -152,8 +158,7 @@ public abstract class ItemUtils {
                 && first.getData().equals(other.getData())
                 && ((!first.hasItemMeta() && !other.hasItemMeta())
                 || (!first.getItemMeta().hasDisplayName() && !other.getItemMeta().hasDisplayName())
-                || (first.getItemMeta().getDisplayName().equals(other.getItemMeta().getDisplayName()))
-            );
+                || (first.getItemMeta().getDisplayName().equals(other.getItemMeta().getDisplayName())));
     }
 
     /**
@@ -326,7 +331,7 @@ public abstract class ItemUtils {
      *
      * @param item An item.
      * @return The Minecraft name of this item, or null if the item's material
-     *      is invalid.
+     *         is invalid.
      * @throws NMSException if the operation cannot be executed.
      */
     public static String getMinecraftId(ItemStack item) throws NMSException {
@@ -407,8 +412,8 @@ public abstract class ItemUtils {
      *
      * @param item An item.
      * @return A NMS ItemStack for this item. If the item was a CraftItemStack,
-     *      this will be the item's handle directly; in the other cases, a copy in a
-     *      NMS ItemStack object.
+     *         this will be the item's handle directly; in the other cases, a copy in a
+     *         NMS ItemStack object.
      * @throws NMSException if the operation cannot be executed.
      */
     public static Object getNMSItemStack(ItemStack item) throws NMSException {
@@ -427,8 +432,8 @@ public abstract class ItemUtils {
      *
      * @param item An item.
      * @return A CraftItemStack for this item. If the item was initially a
-     *      CraftItemStack, it is returned directly. In the other cases, a copy in a
-     *      new CraftItemStack will be returned.
+     *         CraftItemStack, it is returned directly. In the other cases,
+     *         a copy in a new CraftItemStack will be returned.
      * @throws NMSException if the operation cannot be executed.
      */
     public static Object getCraftItemStack(ItemStack item) throws NMSException {
@@ -517,4 +522,94 @@ public abstract class ItemUtils {
         RunTask.nextTick(() -> drop(location, item));
     }
 
+    /**
+     * Converts a chat color to its dye equivalent.
+     *
+     * <p>The transformation is not perfect as there is no 1:1
+     * correspondence between dyes and chat colors.</p>
+     *
+     * @param color The chat color.
+     * @return The corresponding dye, or an empty value if none match (e.g. for formatting codes, of for {@code null}).
+     */
+    @Contract(pure = true)
+    public static Optional<DyeColor> asDye(@Nullable final ChatColor color) {
+        if (color == null) {
+            return Optional.empty();
+        }
+
+        switch (color) {
+            case BLACK:
+                return Optional.of(DyeColor.BLACK);
+
+            case BLUE:
+            case DARK_BLUE:
+                return Optional.of(DyeColor.BLUE);
+
+            case DARK_GREEN:
+                return Optional.of(DyeColor.GREEN);
+
+            case DARK_AQUA:
+                return Optional.of(DyeColor.CYAN);
+
+            case DARK_RED:
+                return Optional.of(DyeColor.RED);
+
+            case DARK_PURPLE:
+                return Optional.of(DyeColor.PURPLE);
+
+            case GOLD:
+            case YELLOW:
+                return Optional.of(DyeColor.YELLOW);
+
+            case GRAY:
+                return Optional.of(DyeColor.LIGHT_GRAY);
+
+            case DARK_GRAY:
+                return Optional.of(DyeColor.GRAY);
+
+            case GREEN:
+                return Optional.of(DyeColor.LIME);
+
+            case AQUA:
+                return Optional.of(DyeColor.LIGHT_BLUE);
+
+            case RED:
+                return Optional.of(DyeColor.ORANGE);
+
+            case LIGHT_PURPLE:
+                return Optional.of(DyeColor.PINK);
+
+            case WHITE:
+                return Optional.of(DyeColor.WHITE);
+
+            // White, reset & formatting
+            default:
+                return Optional.empty();
+        }
+    }
+
+    /**
+     * Converts a dye color to a dyeable material.
+     *
+     * @param material The colorable material to colorize.
+     * @param color    The dye color.
+     * @return The corresponding material.
+     */
+    @Contract(pure = true)
+    public static Material colorize(@NotNull final ColorableMaterial material, @NotNull final DyeColor color) {
+        return Material.valueOf(color.name() + "_" + material.name());
+    }
+
+    /**
+     * Converts a chat color to a dyeable material.
+     *
+     * @param material The colorable material to colorize.
+     * @param color    The chat color.
+     * @return The corresponding material. If the chat color was not convertible to a dye, {@code ChatColor#WHITE} is
+     *         used.
+     */
+    @Contract(pure = true)
+    public static Material colorize(@NotNull final ColorableMaterial material, @NotNull final ChatColor color) {
+        return colorize(material, asDye(color).orElse(DyeColor.WHITE));
+    }
 }